# BookingMovieTicket_API/swagger.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "BookingMovieTicket API",
                 "version":  "v1"
             },
    "paths":  {
                  "/api/Auth/register":  {
                                             "post":  {
                                                          "tags":  [
                                                                       "Auth"
                                                                   ],
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/RegisterRequest"
                                                                                                                                  }
                                                                                                                   },
                                                                                              "text/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/RegisterRequest"
                                                                                                                           }
                                                                                                            },
                                                                                              "application/*+json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/RegisterRequest"
                                                                                                                                    }
                                                                                                                     }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "text/plain":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                      }
                                                                                                                       },
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                            }
                                                                                                                             },
                                                                                                        "text/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                     }
                                                                                                                      }
                                                                                                    }
                                                                                    },
                                                                            "400":  {
                                                                                        "description":  "Bad Request",
                                                                                        "content":  {
                                                                                                        "text/plain":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                      }
                                                                                                                       },
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                            }
                                                                                                                             },
                                                                                                        "text/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                     }
                                                                                                                      }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/Auth/login":  {
                                          "post":  {
                                                       "tags":  [
                                                                    "Auth"
                                                                ],
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/LoginRequest"
                                                                                                                               }
                                                                                                                },
                                                                                           "text/json":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/LoginRequest"
                                                                                                                        }
                                                                                                         },
                                                                                           "application/*+json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/LoginRequest"
                                                                                                                                 }
                                                                                                                  }
                                                                                       }
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "text/plain":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                   }
                                                                                                                    },
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                         }
                                                                                                                          },
                                                                                                     "text/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                 }
                                                                                 },
                                                                         "401":  {
                                                                                     "description":  "Unauthorized",
                                                                                     "content":  {
                                                                                                     "text/plain":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                   }
                                                                                                                    },
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                         }
                                                                                                                          },
                                                                                                     "text/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/StringApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   }
                                      },
...
                  "/api/Auth/profile":  {
                                            "get":  {
                                                        "tags":  [
                                                                     "Auth"
                                                                 ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK"
                                                                                  },
                                                                          "401":  {
                                                                                      "description":  "Unauthorized",
                                                                                      "content":  {
                                                                                                      "text/plain":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                    }
                                                                                                                     },
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                          }
                                                                                                                           },
                                                                                                      "text/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                   }
                                                                                                                    }
                                                                                                  }
                                                                                  }
                                                                      },
                                                        "security":  [
                                                                         {

                                                                         }
                                                                     ]
                                                    }
                                        },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "LoginRequest":  {
                                                            "required":  [
                                                                             "password",
                                                                             "username"
                                                                         ],
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "username":  {
                                                                                                "minLength":  1,
                                                                                                "type":  "string"
                                                                                            },
                                                                               "password":  {
                                                                                                "minLength":  1,
                                                                                                "type":  "string"
                                                                                            }
                                                                           },
                                                            "additionalProperties":  false
                                                        },
...
                                       "RegisterRequest":  {
                                                               "required":  [
                                                                                "email",
                                                                                "fullName",
                                                                                "password",
                                                                                "username"
                                                                            ],
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "username":  {
                                                                                                   "minLength":  3,
                                                                                                   "type":  "string"
                                                                                               },
                                                                                  "email":  {
                                                                                                "minLength":  1,
                                                                                                "type":  "string",
                                                                                                "format":  "email"
                                                                                            },
                                                                                  "password":  {
                                                                                                   "minLength":  6,
                                                                                                   "type":  "string"
                                                                                               },
                                                                                  "fullName":  {
                                                                                                   "minLength":  1,
                                                                                                   "type":  "string"
                                                                                               }
                                                                              },
                                                               "additionalProperties":  false
                                                           },
...
                                       "StringApiResponse":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "success":  {
                                                                                                    "type":  "boolean"
                                                                                                },
                                                                                    "message":  {
                                                                                                    "type":  "string",
                                                                                                    "nullable":  true
                                                                                                },
                                                                                    "data":  {
                                                                                                 "type":  "string",
                                                                                                 "nullable":  true
                                                                                             },
                                                                                    "errors":  {
                                                                                                   "type":  "array",
                                                                                                   "items":  {
                                                                                                                 "type":  "string"
                                                                                                             },
                                                                                                   "nullable":  true
                                                                                               }
                                                                                },
                                                                 "additionalProperties":  false
                                                             },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Dán trực tiếp chuỗi JWT Token vào đây (Swagger sẽ tự động thêm \u0027Bearer \u0027)",
                                                              "scheme":  "Bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {

                     }
                 ],
...
}